Refactor authentication service to use DAO and improve login logic
