+++ src/service/AuthService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace Service;
 
 /**
  * AuthService Interface
