+++ src/service/UserService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace Service;
 
 /**
  * UserService Interface
